fix: structure of project
